Adding a Tutorial page for charting.
As part of one of the main plans for PRSZ, I am
currently working on a tutorial. I have already
made a few folders to accomodate this along with
the first index page. I am currently aware of other
websites which are already doing this, but their
information is scattered all ove the place. This
phase of my ambition aims to bring all this info
together.